increase compilation level of warning (i.e. add -W -Wall) + fix compilation warnings
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/morphologist/morphologist-private/trunk@82629 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ vip/src/VipFileDump/fileDump.c
@@ -90,7 +90,7 @@ int main(int argc, char *argv[])
   size = fileStat.st_size;
   /* ## does not support large files! */
   fileptr = (unsigned char *)calloc((size_t)size, sizeof(unsigned char));
-  if ( fread((void *)fileptr, sizeof(unsigned char), (size_t)size, theFile) != size )
+  if ( fread((void *)fileptr, sizeof(unsigned char), (size_t)size, theFile) != (size_t)size )
   {
     printf("Error reading file...\n");
     fclose(theFile);
